move FilterCheckbox state up to FilterDrawer
